[FIX] #1740
